Radar: move overlay into the dashboard dial gutters
The fullscreen edge lane bar was overkill ,  it pulled the rider's
peripheral attention across every screen even though the radar is
only really actionable while looking at the speedo. Replace it with
two compact mini lanes mounted directly inside the dashboard's dial
Box:

  LEFT  ,  CenterStart, sits in the gutter between the P/D column
            (TopStart) and the Navigator glyph (BottomStart).
  RIGHT ,  CenterEnd,   sits in the gutter between the GPS icon
            (TopEnd) and the Studio camera glyph (BottomEnd).
  BOTH  ,  renders both, for riders who want symmetric coverage.

The mini reuses the lane-bar visual at 22 x 110 dp (vs 26 x ~screen):
rider donut at the bottom, threat dots positioned by distance, soft
red glow on FAST_APPROACH, distance numeral on the closest target,
battery tick across the top. Verified on emulator-5558 with a
synthetic 3-car frame ,  the mini reads cleanly without colliding
with the speedometer arc or the corner glyphs.

Deletes the old fullscreen RadarOverlay.kt and its mount in
MainActivity. RadarOverlayViewModel stays as-is and is shared by
both side mounts so the show/decay/side state lives in one place.

Strings:
  - radar_overlay_both (new, 15 locales)
  - radar_show_overlay_desc updated everywhere ("Only visible on the
    dashboard" instead of "Stays on every screen")
