+++ app/src/main/java/com/eried/eucplanet/MainActivity.kt
@@ -286,12 +286,12 @@ class MainActivity : AppCompatActivity() {
                             },
                             suppressOnPhone = onMapScreen
                         )
-                        // Radar threat lane bar. Sits on a screen edge so it
-                        // doesn't fight the navigation popup (centred) for
-                        // visual space. Its view model gates visibility off
-                        // pairing + connection state, no extra suppression
-                        // needed per screen.
-                        com.eried.eucplanet.ui.radar.RadarOverlay()
+                        // Radar mini lane lives inside DashboardScreen now,
+                        // mounted directly in the dial Box. Keeping the radar
+                        // visible only while the rider is on the dashboard
+                        // matches where their eyes already are and avoids
+                        // fighting other screens (navigator, studio, settings)
+                        // for the screen-edge gutter.
                     }
                 }
             }
